+++ src/lib/utils.ts
@@ -79,3 +79,16 @@ export const timeFormatRegExp = /^([01]\d|2[0-3]):([0-5]\d)$/;
  * RUC format RegExp 12345678-1 [7-9 digits]-[1-9]
  */
 export const rucFormatRegExp = /^\d{7,9}-[1-9]$/;
+
+/**
+ * Normaliza un texto eliminando acentos y convirtiéndolo a minúsculas
+ * @param text Texto a normalizar
+ * @returns Texto normalizado sin acentos y en minúsculas
+ */
+export function normalizeText(text: string): string {
+  return text
+    .normalize("NFD") // Descompone los caracteres acentuados
+    .replace(/[\u0300-\u036f]/g, "") // Elimina los diacríticos (acentos)
+    .toLowerCase() // Convierte a minúsculas
+    .trim() // Elimina espacios en blanco al inicio y final
+}
